+++ demo/src/main/java/com/verygoodsecurity/vgscheckout/demo/payment/PaymentCheckoutActivity.kt
@@ -1,4 +1,4 @@
-package com.verygoodsecurity.vgscheckout.demo.add
+package com.verygoodsecurity.vgscheckout.demo.payment
 
 import android.os.Bundle
 import android.util.Log
